Add ViewCommand class
There is no class available in the app that handles the view command.

Users need to view the latest entries within different categories to
better manage their spending.

A ViewCommand Class is added into the Command package along with 2
methods (execute and getLatestEntries)

Since the View command also needs an execute command, it inherits from
the command class. In order to improve code readability of execute
function, the getLatestEntries returns the latest number of entries
specified by the user. These entries are then parsed into the ui method
to be printed.