UI class: add method to print entries for the user to view
There is no method in UI that prints the list of relevant entries as
requested by the user.

It is not feasible to use current UI methods since they are meant for
other commands.

Adding the printEntriesToBeViewed and formatViewEntries methods ensure
that Entry objects are first formatted into Strings before being printed
out to the user.